+++ src/main/java/roomescape/application/member/dto/MemberResult.java
@@ -1,4 +1,4 @@
-package roomescape.application.member;
+package roomescape.application.member.dto;
 
 import roomescape.domain.member.Member;
