+++ src/main/java/roomescape/presentation/api/member/MemberController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.member.MemberService;
-import roomescape.application.member.MemberResult;
+import roomescape.application.member.dto.MemberResult;
 
 @RestController
 @RequestMapping("/members")
